+++ packages/ui/src/components/separator/separator.tsx
@@ -24,11 +24,8 @@ export const Separator = forwardRef<HTMLDivElement, SeparatorProps>(
         ref={ref}
         orientation={orientation}
         decorative={decorative}
-        className={cn(
-          "shrink-0 bg-border",
-          orientation === "horizontal" ? "h-px w-full" : "w-px self-stretch",
-          className,
-        )}
+        // prettier-ignore
+        className={cn("shrink-0 bg-border", orientation === "horizontal" ? "h-px w-full" : "w-px self-stretch", className)}
         {...props}
       />
     );